+++ resources/views/components/toolbar/actions-toolbar.blade.php
@@ -1,51 +1,32 @@
-{{---------------------------------------------------------------------------
-    Reusable Toolbar component used with edit-create blade template files
+{{-- ------------------------------------------------------------------------
+| -- $routePrefix, is the route name by convention, without the action
+| -- this form has access to the livewire $editing property
+------------------------------------------------------------------------ --}}
 
-    usage:
-    <x-media-toolbar form-name="document-form" />
+@aware(['routePrefix', 'editing'])
 
-    Hide preview button
-    <x-media-toolbar formName="page-form" :preview=false />
+@php
+    $resource = dotLastSegment($routePrefix)
+@endphp
 
+<div id="actions-toolbar" class="pxy-05 light my flex space-between">
 
-    requires redirectById() helper
----------------------------------------------------------------------------}}
+    <div>
 
+        <button wire:click.prevent="save('save_stay');" class="btn sm primary txt-upper">
+            <x-gt-icon-save class="icon" /> <span>Save</span> </button>
 
-<div id="actions-toolbar" class="pxy-05 light flex">
-    <div class="fg1">
-        <button type="submit" form="{{ $formName }}" name="action" value="save_stay" class="btn success sm mr-025">
-            <svg class="icon">
-                <use xlink:href="/svg/naykel-ui-SVG-sprite.svg#save"></use>
-            </svg>
-            <span>SAVE</span>
-        </button>
-        <button type="submit" form="{{ $formName }}" name="action" value="save_close" class="btn success sm mr-025">
-            <svg class="icon">
-                <use xlink:href="/svg/naykel-ui-SVG-sprite.svg#exit"></use>
-            </svg>
-            <span>SAVE & CLOSE</span>
-        </button>
-        <button type="submit" form="{{ $formName }}" name="action" value="save_new" class="btn info sm mr-025">
-            <svg class="icon">
-                <use xlink:href="/svg/naykel-ui-SVG-sprite.svg#plus"></use>
-            </svg>
-            <span>SAVE & NEW</span>
-        </button>
+        <button wire:click.prevent="save('save_close')" class="btn sm primary txt-upper">
+            <x-gt-icon-exit class="icon" /> <span>Save and Close</span> </button>
 
-        <button type="submit" form="{{ $formName }}" name="action" value="save_preview" class="btn secondary sm" {{ $preview ? null : 'disabled' }}>
-            <svg class="icon">
-                <use xlink:href="/svg/naykel-ui-SVG-sprite.svg#preview"></use>
-            </svg>
-            <span>PREVIEW</span>
-        </button>
     </div>
-    <div class="right">
-        <a href="{{ route("$routeName.index") }}" class="btn danger sm">
-            <svg class="icon">
-                <use xlink:href="/svg/naykel-ui-SVG-sprite.svg#back"></use>
-            </svg>
-            <span>BACK TO LIST</span>
+
+    <div>
+
+        <a href="{{ route("$routePrefix.index") }}" class="btn sm txt-upper dark">
+            <x-gt-icon-exit class="icon" /> <span>{{ $resource }} Table</span>
         </a>
+
     </div>
+
 </div>